pyproject.toml
[project]
name = "polymarket_ai_trader"
version = "0.1.0"
description = "AI agents for Polymarket trading."
authors = [{ name = "Tymon Becella", email = "tymon.becella@gmail.com" }]
requires-python = ">=3.10,<4.0"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "pydantic",
    "pydantic-settings",
    "psycopg2-binary",
    "requests",
    "tenacity",
    "ratelimit",
    "supabase",
]

[project.optional-dependencies]
dev = ["pytest", "ruff"]

[tool.ruff]
line-length = 110

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "PL"]

[tool.ruff.format]
quote-style = "double"
